Rename test pkg to testutil
