+++ pkg/skaffold/docker/image_test.go
@@ -22,7 +22,7 @@ import (
 	"os"
 	"testing"
 
-	testutil "github.com/GoogleCloudPlatform/skaffold/test"
+	"github.com/GoogleCloudPlatform/skaffold/testutil"
 )
 
 func TestMain(m *testing.M) {
