+++ pkg/skaffold/docker/parse_test.go
@@ -23,7 +23,7 @@ import (
 
 	"sort"
 
-	testutil "github.com/GoogleCloudPlatform/skaffold/test"
+	"github.com/GoogleCloudPlatform/skaffold/testutil"
 	"github.com/spf13/afero"
 )
